+++ cookbooks/chub-csdashboard/metadata.rb
@@ -6,6 +6,8 @@ description      'Installs/Configures chub-missioncontrol'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends 'java'
+depends 'chub_java'
 depends 'tomcat'
 depends 'openssl'
+depends 'ulimit'
+depends 'mongodb'