+++ roles/elasticsearch.rb
@@ -1,3 +1,16 @@
 name "elasticsearch"
 description "an elasticsearch cluster node"
-run_list "recipe[java]", "recipe[elasticsearch::search_discovery]", "recipe[elasticsearch]"
+run_list "role[base]", "recipe[java]", "recipe[elasticsearch::search_discovery]", "recipe[elasticsearch]"
+default_attributes(
+  "java" => {
+    "install_flavor" => "oracle",
+    "jdk_version" => "8",
+    "jdk" => {
+      "8" => {
+        "x86_64" => {
+          "url" => "http://artifactory01.nexus.commercehub.com/artifactory/java-distributions/jdk/8-b132/jdk-8-linux-x64.tar.gz"
+        }
+      }
+    }
+  })
+