+++ roles/pipeline_team.rb
@@ -4,5 +4,12 @@ default_attributes "sensu" => {
     "rabbitmq" => {
         "host" => "sensu01.nexus.commercehub.com"
     }
+},
+"ssh_keys" => {
+        "lzarou" => "lzarou",
+        "wgrunenwald" => "wgrunenwald",
+        "kstahl" => "kstahl",
+        "bcarr" => "bcarr",
+        "dcross" => "dcross"
 }
-run_list "recipe[chub_pipeline::sensu_client]"
+run_list "recipe[chub_pipeline::sensu_client]","recipe[base::users]","ssh-keys"